+++ docs/CHANGELOG.md
@@ -61,6 +61,10 @@ title: Changelog
 
     *Yoshiyuki Hirano*
 
+* Deprecate engine loading manually.
+
+    *Yoshiyuki Hirano*
+
 ## 2.44.0
 
 * Rename internal accessor to use private naming.